+++ play-frontend-hmrc-play-30/src/main/scala/uk/gov/hmrc/hmrcfrontend/controllers/LanguageController.scala
@@ -20,19 +20,18 @@ import javax.inject.{Inject, Singleton}
 import play.api.Configuration
 import play.api.i18n.Lang
 import play.api.mvc._
-import uk.gov.hmrc.hmrcfrontend.config.LanguageConfig
+import uk.gov.hmrc.hmrcfrontend.config.SupportedLanguagesConfig
 import uk.gov.hmrc.play.language.{LanguageController => PlayLanguageController, LanguageUtils}
 
 @Singleton
 case class LanguageController @Inject() (
   configuration: Configuration,
   languageUtils: LanguageUtils,
-  cc: ControllerComponents,
-  languageConfig: LanguageConfig
+  cc: ControllerComponents
 ) extends PlayLanguageController(languageUtils, cc) {
-  import languageConfig._
+  import SupportedLanguagesConfig._
 
-  override def fallbackURL: String = languageConfig.fallbackURL
+  override def fallbackURL: String = configuration.get[String]("language.fallback.url")
 
   override protected def languageMap: Map[String, Lang] =
     Map(en -> Lang(en), cy -> Lang(cy))